+++ src/main/java/pe/edu/upc/greenly/entities/Rol.java
@@ -8,7 +8,7 @@ public class Rol {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String rol;
 
@@ -21,17 +21,17 @@ public class Rol {
     public Rol() {}
 
     // Constructor con parámetros
-    public Rol(int id, String rol) {
+    public Rol(Long id, String rol) {
         this.id = id;
         this.rol = rol;
     }
 
     // Getters y Setters
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
