+++ src/main/java/pe/edu/upc/greenly/repositories/Ubicacion_CampañaRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.greenly.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.greenly.entities.Ubicacion_Campaña;
 
-public interface Ubicacion_CampañaRepository extends JpaRepository <Ubicacion_Campaña,Integer> {
+public interface Ubicacion_CampañaRepository extends JpaRepository <Ubicacion_Campaña,Long> {
 }
