+++ src/main/java/pe/edu/upc/greenly/repositories/UsuarioRepository.java
@@ -3,5 +3,5 @@ package pe.edu.upc.greenly.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pe.edu.upc.greenly.entities.Usuario;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 }
